Applying authorization

+++ src/main/java/com/practice/spring/airbnb/services/impl/HotelServiceImpl.java
@@ -1,7 +1,9 @@
 package com.practice.spring.airbnb.services.impl;
+
 import java.util.List;
 
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.ResourceAccessException;
@@ -11,6 +13,8 @@ import com.practice.spring.airbnb.dto.HotelInfoDto;
 import com.practice.spring.airbnb.dto.RoomDto;
 import com.practice.spring.airbnb.entities.Hotel;
 import com.practice.spring.airbnb.entities.Room;
+import com.practice.spring.airbnb.entities.User;
+import com.practice.spring.airbnb.exception.UnAuthorizeException;
 import com.practice.spring.airbnb.repositories.HotelRepository;
 import com.practice.spring.airbnb.repositories.RoomRepository;
 import com.practice.spring.airbnb.services.HotelService;
@@ -22,35 +26,39 @@ import lombok.extern.slf4j.Slf4j;
 @Service
 @Slf4j
 @RequiredArgsConstructor
-public class HotelServiceImpl implements HotelService{
+public class HotelServiceImpl implements HotelService {
 
     private final HotelRepository hotelRepository;
     private final ModelMapper modelMapper;
     private final InventoryService inventoryService;
     private final RoomRepository roomRepository;
 
-
     @Override
     public HotelDto cretaeNewHotel(HotelDto hotelDto) {
-       Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
-       hotel.setActive(false);
-       hotel=hotelRepository.save(hotel);
-        return modelMapper.map(hotel,HotelDto.class);
+        Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setActive(false);
+        hotel.setOwner(user);
+        hotel = hotelRepository.save(hotel);
+        return modelMapper.map(hotel, HotelDto.class);
     }
 
     @Override
     public HotelDto getHotelById(Long id) {
         Hotel hotel = hotelRepository.findById(id)
-                      .orElseThrow(() -> new ResourceAccessException("Hotel not found with id "+id));
-        
-                      
-        return modelMapper.map(hotel,HotelDto.class);
+                .orElseThrow(() -> new ResourceAccessException("Hotel not found with id " + id));
+
+        return modelMapper.map(hotel, HotelDto.class);
     }
 
     @Override
     public HotelDto updateHotelById(Long id, HotelDto hotelDto) {
         Hotel hotel = hotelRepository.findById(id)
-                      .orElseThrow(() -> new ResourceAccessException("Hotel not found with id "+id));
+                .orElseThrow(() -> new ResourceAccessException("Hotel not found with id " + id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorizeException("User has no access to update hotel with id " + id);
+        }
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
         Hotel updateHotel = hotelRepository.save(hotel);
@@ -62,25 +70,33 @@ public class HotelServiceImpl implements HotelService{
     @Transactional
     public void deleteHotelById(Long id) {
         Hotel hotel = hotelRepository.findById(id)
-                      .orElseThrow(() -> new ResourceAccessException("Hotel not found with id "+id));
+                .orElseThrow(() -> new ResourceAccessException("Hotel not found with id " + id));
 
-        for(Room room : hotel.getRooms()){
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorizeException("User has no access to delete hotel with id " + id);
+        }
+        for (Room room : hotel.getRooms()) {
             inventoryService.deleteAllInventories(room);
             roomRepository.deleteById(room.getId());
-            
+
         }
         hotelRepository.delete(hotel);
-        
+
     }
 
     @Override
     @Transactional
     public void activateHotel(Long id) {
         Hotel hotel = hotelRepository.findById(id)
-                      .orElseThrow(() -> new ResourceAccessException("Hotel not found with id "+id));
+                .orElseThrow(() -> new ResourceAccessException("Hotel not found with id " + id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorizeException("User has no access to active hotel with id " + id);
+        }
         hotel.setActive(true);
 
-        for(Room room : hotel.getRooms()){
+        for (Room room : hotel.getRooms()) {
             inventoryService.initializeRoomForAYear(room);
         }
     }
@@ -88,15 +104,14 @@ public class HotelServiceImpl implements HotelService{
     @Override
     public HotelInfoDto getHotelInfoById(Long hotelId) {
         Hotel hotel = hotelRepository.findById(hotelId)
-                      .orElseThrow(() -> new ResourceAccessException("Hotel not found with id "+hotelId));
-        
-        List<RoomDto> rooms=hotel.getRooms()
-                            .stream()
-                            .map((elements) -> modelMapper.map(elements,RoomDto.class))
-                            .toList();
-        return new HotelInfoDto(modelMapper.map(hotel,HotelDto.class), rooms);       
+                .orElseThrow(() -> new ResourceAccessException("Hotel not found with id " + hotelId));
+    
+        List<RoomDto> rooms = hotel.getRooms()
+                .stream()
+                .map((elements) -> modelMapper.map(elements, RoomDto.class))
+                .toList();
+        return new HotelInfoDto(modelMapper.map(hotel, HotelDto.class), rooms);
 
     }
-    
 
 }
